3785 - clean up code
