+++ src/client/components/Chart/Bar/Tooltip/Tooltip.tsx
@@ -1,0 +1,25 @@
+import React from 'react'
+
+import { TooltipProps } from 'recharts'
+
+import TooltipContent from 'client/components/Chart/TooltipContent'
+
+const Tooltip: React.FC<TooltipProps<never, never>> = (props) => {
+  const { payload } = props
+
+  if (!(payload.length > 0)) {
+    return null
+  }
+
+  const content = payload.map((item) => ({
+    color: item.color,
+    label: item.name,
+    name: item.name,
+    unit: item.unit,
+    value: item.value as number,
+  }))
+
+  return <TooltipContent content={content} />
+}
+
+export default Tooltip
