+++ src/client/components/Chart/Pie/Tooltip/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Tooltip'
